♻️ [Refactor]: 앱 디렉토리에 있는 컴포넌트 features 디렉토리로 이동
