+++ src/features/studies/components/EditorPreview.tsx
