+++ src/features/studies/components/StudyCard.tsx
