+++ src/features/studies/components/StudyCurriculumCard.tsx
