+++ src/features/studies/components/StudyDescriptionCard.tsx
