+++ src/features/studies/components/applyStudy/ApplyFormCard.tsx
